fix(tests): Move tests to dedicated namespace
Move tests to dedicated namespace so that they don't interfere with others.
